# streams: Vastly increase new stream modal performance.