--- a/static/templates/new_stream_users.handlebars
+++ b/static/templates/new_stream_users.handlebars
@@ -27,12 +27,9 @@
     <a href="#" draggable="false" class="subs_unset_all_users">{{t "Uncheck all" }}</a>
 </div>
 
-<div id="user-checkboxes">
-    {{#each users}}
-    <label class="checkbox add-user-label" data-user-id="{{this.user_id}}" data-email="{{this.email}}">
-        <input type="checkbox" name="user" value="{{this.email}}" />
-        <span></span>
-        {{this.full_name}} ({{this.email}})
-    </label>
-    {{/each}}
+<div class="progressive-table-wrapper grey-box">
+    <table class="table table-condensed table-striped wrapped-table">
+        <tbody id="user-checkboxes" class="admin_user_table required-text thick"
+               data-empty="{{t 'No users found.' }}"></tbody>
+    </table>
 </div>