This changes the user list to render progressively rather than to
render all at once (in chat.zulip.org, that’s 4000+ people who are
rendered!).

This decreases the time to render from a linear speed of render of
around 0.1ms + 4ms to a max of 8ms => means CZO render time goes from
around 400ms => 8ms.

In practice, this actually cuts down on more than a second compared to
CZO.

We also now track whether users are checked in a different way. We make
a single user list from `people.get_rest_of_realm()`, that we add a
custom `__meta` attribute to that tracks whether a user is checked. Now
instead of querying the DOM (which wouldn’t work anymore since it’s a
partial list), we check the attribute.

All said and done, without the duplicated functions and such, a realm with 4000 users will now take around 40ms from start to finish to render the screen rather than 3s. :)

**Orthogonal Performance Benefits:**

Aside from the direct cost of loading, the scrolling in the streams modal is now up back close to 60fps again rather than around 8fps on my computer due to the huge amount of DOM nodes not being rendered anymore.