@brockwhittaker 2017-12-05
**Before Optimization:** <img width="841" alt="Before Optimization" src="https://user-images.githubusercontent.com/10321399/33633192-cc36ea1c-d9c4-11e7-831c-48a66443d830.png">
**After Optimization:** <img width="893" alt="After Optimization" src="https://user-images.githubusercontent.com/10321399/33633193-cc62f7ec-d9c4-11e7-9796-bece805e69ec.png">