@brockwhittaker 2017-12-05
Here's live CZO vs. the new implementation. Keep in mind there are three values (in ms) for the one deployed live because the same function is run three (!!) times.

**Before Optimization:** <img width="1680" alt="screen shot 2017-12-05 at 2 28 33 pm" src="https://user-images.githubusercontent.com/10321399/33634345-b25d1ec8-d9c8-11e7-8383-cf4c97d138df.png">

**After Optimization:** <img width="1680" alt="screen shot 2017-12-05 at 2 28 10 pm" src="https://user-images.githubusercontent.com/10321399/33634346-b2f693b4-d9c8-11e7-8bf8-3a8444ddacba.png">