@showell 2017-12-05
This is a pretty awesome performance increase.  The first three commits look good.  I haven't looked at the fourth one in detail, but it seems reasonable.  There's a Casper test failing, apparently.