@timabbott 2018-02-10
I merged the first 2 commits.  The third breaks tests in a way that suggests it might break some other features.  It looks like the rest needs a rebase.