@timabbott 2020-05-06
Closing as the useful parts of this have been merged; we have a different plan forward for the remainder of the subscribers section.